Unmarshal JSON as rawMessage to be indexed
To reduce memory consumption, input JSON is decoded as a stream to
json.rawMessage instead of a Comic struct. The array of rawMessage is
then indexed directly.

This has also removed the in memory map of all comics in Server. As
such, the results of the search query must be casted before being
unmarshaled.